+++ rides-android/src/main/java/com/uber/sdk/android/rides/RideParameters.java
@@ -201,7 +201,7 @@ public class RideParameters implements Parcelable {
     }
 
     /**
-     * Sets the user agent, describing where this {@link RequestDeeplink} came from for analytics.
+     * Sets the user agent, describing where this {@link RideRequestDeeplink} came from for analytics.
      * @param userAgent to set
      */
     void setUserAgent(@NonNull String userAgent) {